requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
regex
pytest
python-dotenv